Added fix for handling of spaces within parenthesis in values.  Also added handling of nested parenthesis in a value so that it doesn't leave the value parenthesis state until all open parenthesis are matched with a close.
Added  a test for the case I ran across this problem where I had a linear-gradient with an RGBA value for the second transition point, but transparent for the first.  The space was dropped after the transparent, but not the rgba value because the closing paren of the RGBA value transitioned back into the value state from value-parenthesis.